Matrix multiplication with APPML

+++ src/core/blas/mkl/gmmklmatmult.c
@@ -101,7 +101,7 @@ const DMatrix<double>&  operator*(const DMatrix<double>& m, const DMatrix<double
 
 	for(int i=0; i<r.getDim1(); i++)
 		for(int j=0; j<r.getDim2(); j++)
-			r[i][j]=(float) work[i+j*r.getDim1()];
+			r[i][j]=(double) work[i+j*r.getDim1()];
 
 	return r;
 }
